requirements.txt
aioredis>=1.3.1,<1.4.0
discord.py>=1.5.1,<1.6.0
lxml>=4.6.2,<4.7.0
pyhocon>=0.3.57,<0.4.0
